Refactor import paths for UserInfo component and Navbar 🧰
